requirements.txt
fastapi
dnspython
pymongo
motor
uvicorn
passlib
python-jose
python-multipart
bcrypt
pillow==8.3.2